// src/app/models/chamado.ts
export interface Chamado{
    id?:                  any;
    dataAbertura?:     string;
    dataFechamento?:   string;
    prioridade:        string;
    status:            string;
    titulo:            string;
    observacao:       string;
    tecnico:           string;
    cliente:           string;
    nomeCliente:       string;
    nomeTecnico:       string;
}